Add utils module and refactor code
